fix: applyLikeBatch 응답 코드 수정

+++ backend/src/main/java/com/yagubogu/like/dto/LikeBatchRequest.java
@@ -8,6 +8,6 @@ public record LikeBatchRequest(
         @Min(0) Long windowStartEpochSec,
         LikeDelta likeDelta
 ) {
-    public record LikeDelta(Long teamId, Integer delta) {
+    public record LikeDelta(Long teamId, Long delta) {
     }
 }
